refactor: Move Metadata-md functionality out of test and import
